Refactoring complex resources
